+++ client/public/assets/css/nft.scss
@@ -44,7 +44,8 @@
 		}
 		.cd_no_nft_message {
 			text-align: center;
-			color: $color27;
+			color: $color37;
+			font-weight: 600;
 		}
 	}
 }
@@ -116,6 +117,7 @@
 		.cd_field_note {
 			color: $color37;
 			font-size: 12px;
+			font-weight: 600;
 		}
 		input {
 			margin-bottom: 5px;
@@ -196,6 +198,7 @@
 	.cd_nft_contract_confirm_notice {
 		font-size: 0.8em;
 		color: $color37;
+		font-weight: 600;
 	}
 }
