+++ client/public/assets/css/stake/table.scss
@@ -21,6 +21,7 @@
 		text-align: center;
 	}
 	.cd_empty_delegation_message {
-		color: $color27;
+		color: $color37;
+		font-weight: 600;
 	}
 }
